backend/db_database_feature/knowledge_base_database_management/requirements.txt
# fastapi
# uvicorn
# python-multipart
# sqlalchemy  # SQL toolkit and ORM
# pyodbc      # Microsoft SQL Server driver (commented out - using PostgreSQL now)
# python-dotenv  # For managing environment variables

fastapi
uvicorn
sqlalchemy
# pyodbc  # Commented out - MSSQL driver not needed
psycopg2-binary  # PostgreSQL driver
python-dotenv
python-multipart


#------------------Phase 3------------

passlib[bcrypt]
python-jose
email-validator
requests